+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/EditionController.java
@@ -1,4 +1,4 @@
-package ro.societateahermes.backendservice.controller;
+package ro.societateahermes.backendservice.controller.controllerImplementation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
